requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
netaddr==0.7.19
pytz==2019.3
rest-framework-generic-relations==2.0.0
sqlparse==0.3.1
